+++ firebase-firestore/src/androidTest/java/com/google/firebase/firestore/testutil/IntegrationTestUtil.java
@@ -18,7 +18,9 @@ import static com.google.firebase.firestore.testutil.TestUtil.map;
 import static com.google.firebase.firestore.util.Util.autoId;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 import android.content.Context;
 import android.os.StrictMode;
@@ -203,7 +205,7 @@ public class IntegrationTestUtil {
 
   @NonNull
   public static DatabaseId testEnvDatabaseId() {
-    return DatabaseId.forProject(provider.projectId());
+    return DatabaseId.forDatabase(provider.projectId(), BuildConfig.TARGET_DATABASE_ID);
   }
 
   public static FirebaseFirestoreSettings newTestSettings() {
@@ -386,6 +388,12 @@ public class IntegrationTestUtil {
     CollectionReference writer = testFirestore().collection(collection.getId());
     writeAllDocs(writer, docs);
     return collection;
+  }
+
+  public static void writeTestDocsOnCollection(
+      CollectionReference collection, Map<String, Map<String, Object>> docs) {
+    CollectionReference writer = testFirestore().collection(collection.getId());
+    writeAllDocs(writer, docs);
   }
 
   public static void writeAllDocs(
@@ -628,6 +636,71 @@ public class IntegrationTestUtil {
     List<String> expected = asList(expectedDocs);
     if (!expected.isEmpty()) {
       assertEquals(expected, querySnapshotToIds(docsFromQuery));
+    }
+  }
+
+  // Asserts that the given query produces the expected result for all of the
+  // following scenarios:
+  // 1. Performing the given query using source=server, compare with expected result and populate
+  // cache.
+  // 2. Performing the given query using source=cache, compare with server result and expected
+  // result.
+  // 3. Using a snapshot listener to raise snapshots from cache and server, compare them with
+  // expected result.
+  public static void assertSDKQueryResultsConsistentWithBackend(
+      Query collection,
+      Query query,
+      Map<String, Map<String, Object>> allData,
+      List<String> expectedDocIds)
+      throws Exception {
+    // Check the cache round trip first to make sure cache is properly populated, otherwise the
+    // snapshot listener below will return partial results from previous
+    // "assertSDKQueryResultsConsistentWithBackend" calls if it is called multiple times in one test
+    checkOnlineAndOfflineResultsMatch(collection, query, expectedDocIds.toArray(new String[0]));
+
+    EventAccumulator<QuerySnapshot> eventAccumulator = new EventAccumulator<>();
+    ListenerRegistration registration =
+        query.addSnapshotListener(MetadataChanges.INCLUDE, eventAccumulator.listener());
+    List<QuerySnapshot> watchSnapshots;
+    try {
+      watchSnapshots = eventAccumulator.await(2);
+    } finally {
+      registration.remove();
+    }
+    assertTrue(watchSnapshots.get(0).getMetadata().isFromCache());
+    verifySnapshot(watchSnapshots.get(0), allData, expectedDocIds);
+    assertFalse(watchSnapshots.get(1).getMetadata().isFromCache());
+    verifySnapshot(watchSnapshots.get(1), allData, expectedDocIds);
+  }
+
+  public static void verifySnapshot(
+      QuerySnapshot snapshot,
+      Map<String, Map<String, Object>> allData,
+      List<String> expectedDocIds) {
+    List<String> snapshotDocIds = querySnapshotToIds(snapshot);
+    assertEquals(
+        String.format(
+            "Did not get the same document size. Expected doc size: %d, Actual doc size: %d ",
+            expectedDocIds.size(), snapshotDocIds.size()),
+        expectedDocIds.size(),
+        snapshotDocIds.size());
+    assertTrue(
+        String.format(
+            "Did not get the expected document IDs. Expected doc IDs: %s, Actual doc IDs: %s ",
+            expectedDocIds, snapshotDocIds),
+        expectedDocIds.equals(snapshotDocIds));
+
+    Map<String, Object> actualDocs = toDataMap(snapshot);
+
+    for (String docId : expectedDocIds) {
+      Map<String, Object> expectedDoc = allData.get(docId);
+      Map<String, Object> actualDoc = (Map<String, Object>) actualDocs.get(docId);
+
+      assertTrue(
+          String.format(
+              "Did not get the expected document content. Expected doc: %s, Actual doc: %s ",
+              expectedDoc, actualDoc),
+          expectedDoc.equals(actualDoc));
     }
   }
 }